+++ podgen/__main__.py
@@ -7,6 +7,9 @@
         <thorben@sjostrom.no>
     :license: FreeBSD and LGPL, see license.* for more details.
 '''
+# Support for Python 2.7
+from __future__ import absolute_import, division, print_function, unicode_literals
+from builtins import *
 
 import sys
 import datetime