+++ podgen/tests/test_podcast.py
@@ -9,13 +9,18 @@
     :license: FreeBSD and LGPL, see license.* for more details.
 """
 
+# Support for Python 2.7
+from __future__ import absolute_import, division, print_function, unicode_literals
+from builtins import *
+from future.utils import raise_from
+
 import unittest
 import warnings
+import locale
 
 from lxml import etree
 import tempfile
 import os
-from future.utils import raise_from
 
 from podgen import NotSupportedByItunesWarning, Person, Category, Podcast
 import podgen.version
@@ -26,6 +31,8 @@ import dateutil.parser
 class TestPodcast(unittest.TestCase):
 
     def setUp(self):
+        self.existing_locale = locale.setlocale(locale.LC_ALL, None)
+        locale.setlocale(locale.LC_ALL, 'C')
 
         fg = Podcast()
 
@@ -36,7 +43,8 @@ class TestPodcast(unittest.TestCase):
 
         self.name = 'Some Testfeed'
 
-        self.author = Person('John Doe', 'john@example.de')
+        # Use character not in ASCII to catch encoding errors
+        self.author = Person('Jon Døll', 'jon@example.com')
 
         self.website = 'http://example.com'
         self.description = 'This is a cool feed!'
@@ -98,6 +106,9 @@ class TestPodcast(unittest.TestCase):
         def noop(*args, **kwargs):
             pass
         warnings.showwarning = noop
+
+    def tearDown(self):
+        locale.setlocale(locale.LC_ALL, self.existing_locale)
 
     def test_constructor(self):
         # Overwrite fg from setup
@@ -167,6 +178,7 @@ class TestPodcast(unittest.TestCase):
         # Keep track of our temporary file and its filename
         filename = None
         file = None
+        encoding = 'UTF-8'
         try:
             # Get our temporary file name
             file = tempfile.NamedTemporaryFile(delete=False)
@@ -174,9 +186,9 @@ class TestPodcast(unittest.TestCase):
             # Close the file; we will just use its name
             file.close()
             # Write the RSS to the file (overwriting it)
-            fg.rss_file(filename=filename, **kwargs)
+            fg.rss_file(filename=filename, encoding=encoding, **kwargs)
             # Read the resulting RSS
-            with open(filename, "r") as myfile:
+            with open(filename, "r", encoding=encoding) as myfile:
                 rssString = myfile.read()
         finally:
             # We don't need the file any longer, so delete it